+++ R/compounds.R
@@ -821,6 +821,7 @@ setMethod("plotVenn", "compounds", function(obj, ..., labels = NULL, vennArgs = 
     ac <- checkmate::makeAssertCollection()
     checkmate::assertList(allCompounds, types = "compounds", min.len = 2, any.missing = FALSE,
                           unique = TRUE, .var.name = "...", add = ac)
+    checkmate::assertCharacter(labels, min.chars = 1, len = length(allCompounds), null.ok = TRUE, add = ac)
     checkmate::assertList(vennArgs, names = "unique", null.ok = TRUE, add = ac)
     checkmate::reportAssertions(ac)
 
@@ -857,6 +858,7 @@ setMethod("plotUpSet", "compounds", function(obj, ..., labels = NULL, nsets = le
     ac <- checkmate::makeAssertCollection()
     checkmate::assertList(allCompounds, types = "compounds", min.len = 2, any.missing = FALSE,
                           unique = TRUE, .var.name = "...", add = ac)
+    checkmate::assertCharacter(labels, min.chars = 1, len = length(allCompounds), null.ok = TRUE, add = ac)
     checkmate::assertList(upsetArgs, names = "unique", null.ok = TRUE, add = ac)
     checkmate::assertCount(nsets, positive = TRUE)
     checkmate::assertCount(nintersects, positive = TRUE, na.ok = TRUE)
@@ -911,7 +913,7 @@ setMethod("consensus", "compounds", function(obj, ..., absMinAbundance = NULL,
                                              relMinAbundance = NULL,
                                              uniqueFrom = NULL, uniqueOuter = FALSE,
                                              minMaxNormalization = FALSE,
-                                             rankWeights = 1)
+                                             rankWeights = 1, labels = NULL)
 {
     allCompounds <- c(list(obj), list(...))
 
@@ -919,10 +921,11 @@ setMethod("consensus", "compounds", function(obj, ..., absMinAbundance = NULL,
     checkmate::assertList(allCompounds, types = "compounds", min.len = 2, any.missing = FALSE,
                           unique = TRUE, .var.name = "...", add = ac)
     checkmate::assertNumeric(rankWeights, lower = 0, finite = TRUE, add = ac)
+    checkmate::assertCharacter(labels, min.chars = 1, len = length(allCompounds), null.ok = TRUE, add = ac)
     checkmate::reportAssertions(ac)
 
     rankWeights <- rep(rankWeights, length.out = length(allCompounds))
-    compNames <- sapply(allCompounds, algorithm)
+    compNames <- if (!is.null(labels)) labels else sapply(allCompounds, algorithm)
     if (anyDuplicated(compNames))
     {
         # duplicate algorithms used, try to form unique names by adding library